src/requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.2
click==7.1.1
Django==3.0.5
django-appconf==1.0.4
django-debug-toolbar==2.2
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.3
pep8==1.7.1
pilkit==2.0
Pillow==7.1.1
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
